Some changes to <javadoc>
(1) remove the title attribute completely

(2) list project names instead of module names in javadoc.html - not
the difference this makes for vindico which used to be listed as
jakarta-alexandria.

(3) add a defined-in attribute for javadocs that attach themselves to
a different module and use that to remove the false alarms for
Antidote, Avalon-Excalibur and others from nojavadoc.html.

Please note, that these are my very first steps in XSLT, if I'm doing
something totally inefficient or bad, please teach me how to do it
better 8-)