Next try - break up xdoclet's build process in several steps to avoid the joy of static caches.
Detailed:  xdoclet searches META-INF/xdoclet.xml to register modules
and stores the modules found in a static cache.  The "externalizer" is
part of the xdoclet module.

In the build file for the xdoclet module META-INF/xdoclet.xml will be
created by an XDoclet task in xdoclet-xml - this task fills the module
cache and says the xdoclet module is not there.

The very next build step of the xdoclet module needs the
"externalizer" which cannot be found.

The XDoclet people never run into the problem as they rely on
<taskdef> to give them separate class loaders for their tasks
(i.e. each task gets a cache of its own).  In Gump, all XDoclet
classes live on the system classloader and thus share the same cache.