Trick xmlunit by giving it some arbitrary directories to chew on.
xmlunit wants to know the parent directories of xalan.jar and
junit.jar to add them to <classpath>s.  As we are in
build.sysclasspath=only mode, it doesn't matter whether the path is
correct, the <classpath> won't get used anyway.

I could have used reference="home" for xalan and junit, but with no
real gain IMHO.