This is me abusing my newly granted CVS rights by tweaking Axis's Gump descriptor, when clearly I have no clue about Gump descriptors.
Axis is breaking on ant1.6 (not ant1.5) with lib.dir being relative to sub-projects, not absolute as expected. So in Axis we set axis.lib.dir to the location ,/lib, then lib.dir to the location of ${axis.lib.dir}, forcing lib.dir to be absolute regardless of how axis.lib.dir was set.

I suspect that <property path=> would do the same, but there my ignorance of Gump would just start making me dangerous.