Bootstrap now compiles - commence nagging.
dIon (or others) it looks like maven is now proceeding to the point
where it wishes to download its own dependencies.  Could you take a look at
providing an option in maven to *not* do this and instead use the ones
that are provided?

Perhaps some combination of maven.bootstrap.online, maven.jar.override,
or other existing property would do the trick.