Here is adams rework.  I moved the cvs log to the bottom, and added a comment about using a local_env.sh file.
Commit notes from Adam Jack:

I've taken your enhancements and tried to go one step further.

I have this accepting three environment variables GUMP/GUMP_WS/GUMP_LOG and
try to do everything off those. It exits (with 1 -- that right?) if they are
not set. I set those three and it seems to run in my environment. Normally
I'd expect folks to hack these entries.

1) I have it copy `hostname -s`.xml (likely workspace) and 'likely profile'
and the gump.sh to myprofile:

	cp $GUMP/gump.sh $GUMP_PROFILE_LOG_DIR
	cp $GUMP_HOST.xml  $GUMP_PROFILE_LOG_DIR
	cp `grep profile $GUMP_HOST.xml  | cut -d\" -f2` $GUMP_PROFILE_LOG_DIR

2) I added some CVS headers, but am not sure I got those right.

3) I set it as bash not sh -- that ok?

4) I made it echo some separators, but it could use more verbose information
about each section.

5) I made the output be gump.html (better name than gen.html).

6) I don't like the way cron is detected (for running nag.pl) something
better is needed.

7) I made it do the CVS update each time, and a build clean each time. Maybe
these are overkill.

This is a major work in progress (it'd be nice if it output HTML as pretty
as build.sh or update.sh) and it needs a few switches (so folks don't have
to do everything every time.)

This is is "Linux" format.


Submitted by:	Adam Jack ajack@TrySybase.com