Make jstl-jsp-12 compile, add saxpath.
It took me what seemed like infinite iterations to get the dist target
to finally build, just to see that it wouldn't create the jars at all.

Switched to library-dist which creates the jars in a strange place but
probably doesn't need the tweaks I had done before, but I couldn't be
bothered to figure out, which changes could be reverted.