Adapt jar names.
This is not a good fix as we are loosing the date information.

Pluto's build system uses a top level build file that delegates to
three build files in the api, container and portal subdirs, all three
of which use the same property to determine the jar name, so we can't
override it.

The real fix would be to split the jakarta-pluto build into three
builds, but I don't have the time to do it properly right now.