Start on more intelligent build algorithm.
It turns out that the 'multicastplugin' (or whatever it is that acts as visitor to the walker) basically is the right place to implement smarter algorithms. I.e. anything above and beyond the depth-first graph walking goes into the visitor. For now I've started by extending multicastplugin. We might wish to rename that, break it out into its own module, and remove it a little from 'public' view. I'm not sure yet.

The code I'm committing doesn't work at the moment (well it seems like it does, it just skips all build steps since there was no module that was updated), doesn't have tests, and should probably be refactored. I'm checking it in so we can have a discussion about the whole multicast business. If I'm right, we'll want several more of those, all implementing different algorithm ideas.

Stefano, if you're reading, this is the place to do the graph magic :-)