Rework and dramatically simply classpath management. See
http://mail-archives.apache.org/mod_mbox/gump-general/200505.mbox/%3c0e0201c55c01f1b1420401a8c0@sybase.com%3e

for background, as well as

http://issues.apache.org/jira/browse/GUMP-117

and 

http://issues.apache.org/jira/browse/GUMP-118


and

http://mail-archives.apache.org/mod_mbox/gump-general/200505.mbox/%3cBEBB728C.3D48E%25mail@leosimons.com%3e
http://mail-archives.apache.org/mod_mbox/gump-general/200505.mbox/%3cBEBB77D8.3D49A%25mail@leosimons.com%3e

for more extensive discussion.

Basically, the old implementation delayed calculation of lots of paths instead of doing it in the objectification stage. Doing it there makes a *lot* of the pain go away. Other changes include simplification of the classpath calculation algorithm (and moving it into a utility method), code formatting, making a Work directory into an Output, getting rid of all the 'artifact' model objects, not swalling exceptions in the plugins, etc. Most of those change see lots of discussion in the linked mailing list threads.